requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
pydantic>=2.5.2
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.3.7
transformers>=4.35.2
torch>=2.1.1
numpy>=1.26.2
pandas>=2.1.3
python-multipart>=0.0.6 